fix: rhythm part refactoring
