--- src/app/components/contents/Rhythm/Game.module.css
@@ -1,85 +1,0 @@
-.gameContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    background-color: #ececec;
-    position: relative;
-}
-
-.successMessage {
-    font-size: 24px;
-    color: green;
-    margin: 10px 0;
-    height: 30px;
-}
-
-.notesContainer {
-    position: relative;
-    width: 300px;
-    height: 400px;
-    background-color: #ececec;
-    overflow: hidden;
-}
-
-.note {
-    position: absolute;
-    width: 50px;
-    height: 18px;
-    border-radius: 4px;
-    cursor: pointer;
-    transition: transform 4s linear;
-}
-
-.clicked {
-    animation: clickEffect 0.5s forwards;
-}
-
-@keyframes clickEffect {
-    0% {
-        transform: scale(1);
-        opacity: 1;
-    }
-    100% {
-        transform: scale(1.5);
-        opacity: 0;
-    }
-}
-
-.reactiveZone {
-    position: absolute;
-    bottom: 0;
-    width: 50px;
-    height: 50px;
-    border: 2px dashed #000;
-    box-sizing: border-box;
-    opacity: 0.2;
-}
-
-.startButton {
-    padding: 20px 40px;
-    font-size: 20px;
-    cursor: pointer;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    z-index: 1;
-}
-
-.endMessage {
-    margin: 10px 0;
-    font-size: 24px;
-    cursor: pointer;
-    color: red;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    z-index: 1;
-}
-
-.score, .missedNotes {
-    margin-top: 20px;
-    font-size: 24px;
-    background-color: cornflowerblue;
-}
