+++ src/app/components/contents/rhythm/Songs.ts
@@ -1,0 +1,49 @@
+const Twinkle_Twinkle = {
+    name:"Twinkle Twinkle",
+    note :[
+        261.63, 261.63, 392.00, 392.00, 440.00, 440.00, 392.00,
+        349.23, 349.23, 329.63, 329.63, 293.66, 293.66, 261.63,
+        392.00, 392.00, 349.23, 349.23, 329.63, 329.63, 293.66,
+        392.00, 392.00, 349.23, 349.23, 329.63, 329.63, 293.66,
+        261.63, 261.63, 392.00, 392.00, 440.00, 440.00, 392.00,
+        349.23, 349.23, 329.63, 329.63, 293.66, 293.66, 261.63
+    ],
+    speed: 500
+};
+
+const Little_lamb = {
+    name: "Mary had a little lamb",
+    note : [
+        329.63, 293.66, 261.63, 293.66, 329.63, 329.63, 329.63,
+        293.66, 293.66, 293.66, 329.63, 392.00, 392.00,
+        329.63, 293.66, 261.63, 293.66, 329.63, 329.63, 329.63,
+        329.63, 293.66, 293.66, 329.63, 293.66, 261.63,
+        329.63, 293.66, 261.63, 293.66, 329.63, 329.63, 329.63,
+        293.66, 293.66, 293.66, 329.63, 392.00, 392.00,
+        329.63, 293.66, 261.63, 293.66, 329.63, 329.63, 329.63,
+        329.63, 293.66, 293.66, 329.63, 293.66, 261.63
+    ],
+    speed:350
+};
+
+const Fur_Elise = {
+    name: "Fur_Elise",
+    note: [
+        659.25, 622.25, 659.25, 622.25, 659.25, 494.88, 587.33, 523.25,
+        440.00, 261.63, 329.63, 440.00, 493.88, 329.63, 415.30, 493.88,
+        523.25, 329.63, 440.00, 493.88, 523.25, 493.88, 440.00, 329.63,
+        659.25, 622.25, 659.25, 622.25, 659.25, 494.88, 587.33, 523.25,
+        440.00, 261.63, 329.63, 440.00, 493.88, 329.63, 415.30, 493.88,
+        523.25, 329.63, 440.00, 493.88, 523.25, 493.88, 440.00, 329.63,
+        659.25, 622.25, 659.25, 622.25, 659.25, 494.88, 587.33, 523.25,
+        440.00, 261.63, 329.63, 440.00, 493.88, 329.63, 415.30, 493.88,
+        523.25, 329.63, 440.00, 493.88, 523.25, 493.88, 440.00, 329.63,
+        659.25, 622.25, 659.25, 622.25, 659.25, 494.88, 587.33, 523.25,
+        440.00, 261.63, 329.63, 440.00, 493.88, 329.63, 415.30, 493.88,
+        523.25, 329.63, 440.00, 493.88, 523.25, 493.88, 440.00, 329.63
+    ],
+    speed: 200
+};
+
+
+export const levelSeq = [Twinkle_Twinkle, Little_lamb, Fur_Elise]
